Rename cli.rs to args.rs
